Fort Collins Event Importer for Google Calendar
An app for scraping various websites for businesses in the Fort Collins
area, gathering information on upcoming events, and importing them into
a Google Calendar to easily see events for specific days/times. Current
importer.py is just the tutorial from the Google website to ensure the
API is working (it is).