Added Event Exists Check
Closes #7
Was able to just search for the title of an event in order to see if it
already exists or not. This is for sure going to run into issues for any
events that have a repeat title, when that happens I can expand it to
also check the date(s) of any matching events, and if the requested
event and calendar event share a day assume it is the same event and
don't add it. That's a future Noah problem.