Added Avogadros File
Started the code to import avogadros. Good news: their website is static
so it is able to be accessed from just requests.get. Bad news: the site
is all over the place with how they name things and are not strict at
all with double checking/coherency. I've been able to seperate the
events pretty well, the annoying bit is going to be getting the
date/time and event name right every time.