Completed Avogadro's Number Import
Closes #4
With all of the events on the main page, was able to get all of them
with the correct name, date, start and end times. Some random edge cases
had to be handled, and may have to be added to in the future as the
website is updated. File still needs commenting, and some error
protection.