Event Exists Check Start Time
Closes #17
When checking if an event exists, it will search for the title like
normal, but in additional will check all returned events to see if their
start times line up. Was able to remove the date from the title of avo's
events. Also added try catch statement to avo's and comedy fort scrapers
in case the connection fails.