Imported common.py into Avogadro's
Closes #19
Added common.py to Avogadros, cleaned up the code a little bit, but it
is still fairly messy. Added an array of "words that appears alongside
times when they shouldn't have" to add to as more events are
added/changed. Avogadro's also wasn't putting dateTime in the correct
format, the API was handling it, but now it is using the common
functions so the formatting is correct.