Added Specific Calendars
Closes #21
The Calendar now has a sub-calendar for each venue that its events
should be uploaded to. Made a function to get that list of calendars and
their ids, and now uploadingEvent requires the calenderId as well. An
issue also came up while doing this, where events being returned while
checking if an event existed had a dateTime with 'Z' at the end to
represent UTC time. This was causing errors to be thrown when converting
it to datetime.datetime, the solution was to replace 'Z' with "-00:00"
for UTC.