Magic Rat Streamlining + Better Browser Handling
Closes #20
Updated Magic Rat Scraper so that it is now using common defaults, and
added datetime so streamline how the endTime is being calculated. Before
it was done with a lot of splitting, which there is no need for,
datetime can handle that kind of addition. Also, in importer.py, browser
was just a variable, updated to be within a with statement so that the
browser will clean up if the program is force stopped/encounters an
error.