Turned importer.py Main into Loop
Closes #26
Rather than having each venue just repeat the code of the venue before
it, a loop is now used with all of the imported files, that runs the
same lines of code for all of them. Only thing that needs to be updated
now is that venues array with any new additions.
All scrapers were also given a global name variable, that is the name of
the venue/sub-calendar.