Created Main Maxline Loop
Got the main loop going that gathers all of the events. Now it is just a
matter of getting the start/end dateTime for events. There is a default
for when the time isn't listed based on if it's Sunday or Thursday, if
the time is different it is listed in the band name. Of course, this
value isn't standard across all events, because why would it be?