Upgraded Magic Rat from 3 to 6 Events
Closes #9
Was able to upgrade the number of events gotten from Magic Rat from 3 to
6, just by changing the url. Unfortunately, trying to load more failed.
I tried using selenium to click the load more button, and it ends up
causing a connection error. I can't figure out what is causing it, will
create a new issue to try and increase the range further.