Start of Event Verifier
Created the function, and have the first checks to ensure all 4 of the
needed keys are present: summary, description, start, and end. If start
or end is not present, the function will return None. Next, need to
ensure that start and end both have dateTime if it makes it through. If
not, also return None. Finally, if it makes it to that point, ensure
dateTime is in the correct format, and then I can remove the try
statement from uploadEvents.