Implemented Event Verification
Closes #30
Added the needed code to importer in order for every event about to be
uploaded to run through the verifier. Ran the code, no issues happened.
All of the events imported have always been in the right format, but
from the tests it should be able to handle the janky events.