Magic Rat/Avogadro's Fixes
Avogadro's and Magic Rat both had errors when I tried running the
program again.

Avogadro's: there was an event that as just April 28 &29, and it was
causing issues. Made it so that if the program can't find the month,
instead of trying to return a default month, return None and check for
that in the main code.

Magic Rat: There was a closure due to a private event being held at the
venue, and because of this the dateTime included null for the time,
which threw an error in the code. Now, the code will attempt to convert
the startTime from isoformat, and if it can't it will skip to the next
event.